refactor(filters): restructure folders
