+++ packages/admin-ui/src/filters/Filter/filter-radio.tsx
@@ -1,6 +1,6 @@
 import { useElement } from '@vtex/admin-ui-react'
 
-import * as style from '../radio/radio.style'
+import * as style from '../../radio/radio.style'
 
 export const FilterRadio = (props: FilterRadioProps) => {
   const { checked } = props
