+++ packages/admin-ui/src/filters/FilterMultiple/filter-multiple-search.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
 
-import { BaseFilter } from './filter-base'
-import { ComboboxItem } from '../combobox/combobox-item'
-import { FilterSeachbox } from './filter-searchbox'
-import { MultipleItemsLabel } from './MultipleItemsLabel'
-import { Box, Checkbox } from '..'
-import { useMessageFormatter } from '../i18n'
-import { messages } from './filter.i18n'
-import type { FilterMultipleProps } from '.'
+import { BaseFilter } from '../filter-base'
+import { ComboboxItem } from '../../combobox/combobox-item'
+import { FilterSeachbox } from '../filter-searchbox'
+import { MultipleItemsLabel } from './filter-multiple-items-label'
+import { Box, Checkbox } from '../..'
+import { useMessageFormatter } from '../../i18n'
+import { messages } from '../filter.i18n'
+import type { FilterMultipleProps } from '..'
 
-import * as style from './filter.style'
+import * as style from '../filter.style'
 
 export function FilterMultipleSearch<T>(props: FilterMultipleProps<T>) {
   const formatMessage = useMessageFormatter(messages.searchBox)
