+++ packages/admin-ui/src/filters/__tests__/filter-multiple.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useFilterMultipleState } from '..'
 import { render, axe, withState } from '../../test-utils'
 
-import { FilterMultiple } from '../filter-multiple'
+import { FilterMultiple } from '../FilterMultiple/filter-multiple'
 
 const FilterWithState = withState(FilterMultiple, () =>
   useFilterMultipleState({
