+++ packages/admin-ui/src/filters/index.tsx
@@ -1,8 +1,8 @@
-export * from './filter-multiple'
-export * from './filter'
+export * from './FilterMultiple/filter-multiple'
+export * from './Filter/filter'
 export * from './filter-group'
 export * from './filter-group.state'
-export * from './filter-multiple.state'
-export * from './filter.state'
-export * from './filter-multiple-search'
-export * from './filter-search'
+export * from './FilterMultiple/filter-multiple.state'
+export * from './Filter/filter.state'
+export * from './FilterMultiple/filter-multiple-search'
+export * from './Filter/filter-search'
